Partially functioning family member search
The basic idea is that you send the URN for the orpha data element in
FamilyMember, and get back either true or false, meaning that there
either is a relative of the patient with the same Orpha code, or there
isn't.

But the actual value of the Orpha code doesn't matter in this query,
because the CQL on the server side ought to be doing the job of
comparing Orpha code between patient and relative.

I have made two attempts to integrate thia into the GUI.

1. Use the existing selector infrastructure.
In addition to the "Donor/Clinical Information" and "Sample" sections of
the GUI, I have added a "Relatives" section. Here you can select just
one field type, namely "DIOGNOSIS ORPHA". You can enter anything at all
into the field, the actual value will be ignored. The important thing is
that something is there. Click "SEND", and a query will be run to
determine if there is a relative for a given patient.

This is tested and working, so hooray!

But it isn't a nice way to do things, it would be really confusing for
users.

2. Use a checkbox
What we have here is really a binary thing: do we want to restrict the
search to patients with relatives or not? For that purpose, a checkbox
is sufficient. I discovered that the detailed results uses checkboxes,
so I copied and pasted some code from there.

Unfortunately, I don't really understand how the query building process
functions under the bonnet, so my attempts up to now have failed, and
while this checkbox control is visible and clickable, it doesn't actually
do anything.