Disabled the MDR_* environment variables
These variables are used to define the content of the GUI.

Due to the way that the scripts have been written for Docker, the
env.json file is completely overwritten, so that the things you
define there are ignored, even if you don't define any of the
MDR_* variables.

By commenting out the offending lines, I have re-enabled the
env.json file, but the MDR_* environment variables no longer
have an effect. This allows Thomas' changes to work.

And defining the GUI in the docker-compose.yml file seems rather
ugly.