Eliminate dependency on mdr16 namespace
When the Sample Locator GUI opens, it provides by default a field
for entering ICD10 diagnosis codes.

To achieve this, the URN for diagnosis was hard-coded to be:

urn:mdr16:dataelement:27:1

This makes the code inflexible, because it only works for the mdr16
namespace.

The patch being submitted here solves this by looking for the ID '27'
in all of the URNs known to the MDR namespace being used, and using
the corresponding URN for diagnosis if found.

This is still not completely generic of course, it still depends on MDR
developers retaining the ID '27' for diagnosis. However, it works for
the two namespaces being used in GBA and BBMRI, namely mdr16 and
bbmri-1-2, see:

https://locator.bbmri-eric.eu/v3/api/mdr/dataelements/urn:mdr16:dataelement:27:1
https://locator.bbmri-eric.eu/v3/api/mdr/dataelements/urn:bbmri-1-2:dataelement:27:1