Activate the "Send" button when the branding UI is "GBA".
Original situation: if the branding UI was "GBA", the "Send" button
was grayed out and inactive. It was not possible to run a search.

This patch enables the button and also sends the query for execution.

The functionality with branding UI "BBMRI" was left unchanged, since
this is working correctly.