Problems with the handling of SILENT_RENEW
I was playing with the latest develop branch of sample-locator and
discovered that the web page was not loading at all. This was with
Firefox 100.0 (64-bit) on Ubuntu 20.04. The web developer tools told
me that there was a problem with a trailing comma in an env.js file.
I tracked this down to docker/config/env.template.js and fixed it.

When I tried the sample-locator with the fix, the page still didn't
load! This time with an "unexpected }" error. I suspected that the
problem was probably due to SILENT_RENEW being undefined, so I
put a default value into the Dockerfile. That fixed the problem:
now the sample-locator page loads OK.

Moot point: I set the default SILENT_RENEW to 'true', because it
seems like a desirable feature to me. What do the reviewers think?

BTW, this might also explain why the new "silent renew" feature works
in some browsers but not others: different browsers and even versions of
browsers might be more forgiving about JSON syntax errors, allowing
the new feature to work in some browsers but not in others. It might
be worth testing the feature again with my fix, to check this
hypothesis.